backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
alembic==1.11.1
pydantic==1.10.8
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
aiofiles==23.1.0
pymysql==1.0.3
email-validator==2.0.0
oss2==2.18.0
httpx==0.24.1
python-docx==0.8.11